==> src/2_mdtrajanalysis_splittraj.ipynb <==
import mdtraj
import tqdm
import os

# Find the trajectory named step7_production_whole_dry_compact_center.xtc in md_sims_norA\2021_09_21_charmmGUI_norA\gromacs, and load it.
traj = mdtraj.load('../md_sims_norA/2021_09_21_charmmGUI_norA/gromacs/step7_production_whole_dry_compact_center.xtc', 
                   top='../md_sims_norA/2021_09_21_charmmGUI_norA/gromacs/step6.6_equilibration_dry.gro')



# Split this traj into multiple frames, and save them as .pdb files.
folder = "2021_09_21_charmmGUI_norA_splitPDB"

# If folder does not exist, create it.
if not os.path.exists(folder):
    os.makedirs(folder)

for i in tqdm.tqdm(range(0, traj.n_frames)):
    traj[i].save_pdb(f'{folder}/frame_{i}.pdb')
