==> src/4_mdtrajanalysis_parse.ipynb <==
import pandas
import tqdm
import os
import glob
from pathlib import Path

# Get a list of all *_desc.txt files with their full path in the directory 2021_09_21_charmmGUI_norA_splitPDB_dss
txt_list = glob.glob('2021_09_21_charmmGUI_norA_splitPDB_dss/*_desc.txt')

# Parse each of these files into a single dataframe
df = pandas.DataFrame()
for i in tqdm.tqdm(range(0,len(txt_list))):
    txt = txt_list[i]
    df2 = pandas.read_csv(txt, sep='\t')

    # Get the basename of the file
    txt2 = os.path.basename(txt)

    # Add a column to the dataframe with the basename of the file
    df2['frame'] = txt2
    df = pandas.concat([df,df2], ignore_index=True)

# Extract the number in frame column and strore it instead in frame column.
df['frame'] = df['frame'].str.extract('(\d+)').astype(int)

df.to_csv('df_pockets_2021_09_21_charmmGUI_norA.csv', index=False)
