==> src/5_mdtrajanalysis_process.ipynb <==
import pandas
import tqdm
import os
import glob
from pathlib import Path
import biotite.structure.io.pdb as pdb
import biotite.database.rcsb as rcsb
import numpy as np
import nglview as nv
import matplotlib

# Read in df_pockets_2021_09_21_charmmGUI_norA.csv
df_pockets = pandas.read_csv('df_pockets_2021_09_21_charmmGUI_norA.csv')
# Remove rows with drugScore values lower than 0.5
df_pockets = df_pockets[df_pockets['drugScore'] > 0.5]
# Summarize df_pockets
df_pockets.describe()

def annot_pocket_res(df_pockets):
    # Iterate over all rows in df_pockets
    for index, row in tqdm.tqdm(df_pockets.iterrows(), total=df_pockets.shape[0]):
        # Get the name and frame
        name = row['name']
        frame = row['frame']

        pocket_pdb = f'frame_{frame}.pdb_res_{name}.pdb'
        
        # Load the pocket PDB file
        pocket = pdb.PDBFile.read(os.path.join('2021_09_21_charmmGUI_norA_splitPDB_dss',pocket_pdb))

        # Find out unique residue numbers
        res_nums = list(map(int,[pocket.get_structure()[0][i].res_id for i in range(0,len(pocket.get_structure()[0]))]))

        # Get unique values in res_nums
        unique_res_nums = np.unique(res_nums)

        unique_res_str = ' '.join(unique_res_nums.astype(str))

        # Add unique_res_str to df_pockets as a new column
        df_pockets.loc[index, 'pocket_res'] = unique_res_str

    return df_pockets

ncores = 20

# Split df_pockets into ncores chunks.
df_pockets_chunks = np.array_split(df_pockets, ncores)

# Call annot_pocket_res() on each chunk using multiprocessing
from multiprocessing import Pool
with Pool(ncores) as p:
    df_pockets = pandas.concat(p.map(annot_pocket_res, df_pockets_chunks))


# Save df_pockets to a CSV file
# ds_05: drugScore > 0.5
# wpr: with pocket residues
df_pockets.to_csv('df_pockets_ds_05_wpr.csv', index=True)


# (Optional) Remove rows that contain more than a single underscore in the name column
df_pockets = df_pockets[~df_pockets['name'].str.contains('_.*_')]

master_res_list = list()
# Iterate over all rows in df_pockets
for index, row in tqdm.tqdm(df_pockets.iterrows(), total=df_pockets.shape[0]):
    # Convert pocket_res to a list containing integers
    pocket_res = list(map(int, row['pocket_res'].split(' ')))

    # Add pocket_res to master_res_list
    master_res_list.extend(pocket_res)

# Get how frequently each value is observed in master_res_list in a data frame
df_res_freq = pandas.DataFrame(pandas.Series(master_res_list).value_counts())

df_res_freq

# Load frame_0_freq.pdb in nglview
view = nv.show_file('2021_09_21_charmmGUI_norA_splitPDB/frame_0.pdb',default_representation=True)

# Add a spacefill representation of the top 20 residues with highest frequency
for res in df_res_freq.index[:20]:

    view.add_spacefill(selection=f'{res}:A',color='blue')

view.center()
view

file_path = rcsb.fetch('1a0j', 'pdb')
stack = pdb.PDBFile.read(file_path)

mol = pdb.get_structure(stack,extra_fields=["b_factor"])

mol[0].b_factor
